{{cookiecutter.project_slug}}/requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.1  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
{%- if cookiecutter.use_compressor == "y" %}
{%- if cookiecutter.windows == 'y' and cookiecutter.use_docker == 'n' %}
rcssmin==1.0.6 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
{%- else %}
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
{%- endif %}
{%- endif %}
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
{%- endif %}
redis==3.5.3  # https://github.com/andymccurdy/redis-py
{%- if cookiecutter.use_docker == "y" or cookiecutter.windows == "n" %}
hiredis==2.0.0  # https://github.com/redis/hiredis-py
{%- endif %}
{%- if cookiecutter.use_celery == "y" %}
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.1.0  # https://github.com/celery/django-celery-beat
django-timezone-field==4.0  # https://github.com/celery/django-celery-beat/pull/378
{%- if cookiecutter.use_docker == 'y' %}
flower==0.9.7  # https://github.com/mher/flower
{%- endif %}
{%- endif %}
{%- if cookiecutter.use_async == 'y' %}
uvicorn[standard]==0.13.0  # https://github.com/encode/uvicorn
{%- endif %}

# Django
# ------------------------------------------------------------------------------
django==3.1.11  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.11.2  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5  # https://github.com/django-crispy-forms/crispy-bootstrap5
{%- if cookiecutter.use_compressor == "y" %}
django-compressor==2.4  # https://github.com/django-compressor/django-compressor
{%- endif %}
django-redis==4.12.1  # https://github.com/jazzband/django-redis
{%- if cookiecutter.use_drf == "y" %}
# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers
{%- endif %}
django-silk
django-postgres-metrics
django-better-admin-arrayfield
django-adminactions
django-admin-export-action
django-permissions-auditor
data-wizard
itertable

fontawesome-free
django-smart-selects==1.5.9 # https://github.com/jazzband/django-smart-selects
django-import-export # https://github.com/django-import-export/django-import-export https://django-import-export.readthedocs.io/en/latest/installation.html
